refactor: move codes to store getters
